refactor: mvc패턴에 맞게 의존성 수정
